CameraUtils struct moved to a separate file, .gitignore improved
1) CameraUtils struct was moved to file camera_utils.h, which is now
included in camera_models_crtp.h. Rationalle: breaking up a very long
header file to make it easier to navigate.

2) .gitignore now contains CMake-generated things and binary names,
making the project more in-source-build-friendly.
